+++ src/mod/cgab.channels.c
@@ -1,13 +1,13 @@
 #include "gab.h"
 
-a_gab_value *gab_lib_close(struct gab_triple gab, size_t argc,
+a_gab_value *gab_lib_close(struct gab_triple gab, uint64_t argc,
                            gab_value argv[argc]) {
   gab_chnclose(gab_arg(0));
 
   return nullptr;
 }
 
-a_gab_value *gab_lib_isclosed(struct gab_triple gab, size_t argc,
+a_gab_value *gab_lib_isclosed(struct gab_triple gab, uint64_t argc,
                               gab_value argv[argc]) {
   bool closed = gab_chnisclosed(gab_arg(0));
 
@@ -16,7 +16,7 @@ a_gab_value *gab_lib_isclosed(struct gab_triple gab, size_t argc,
   return nullptr;
 }
 
-a_gab_value *gab_lib_isfull(struct gab_triple gab, size_t argc,
+a_gab_value *gab_lib_isfull(struct gab_triple gab, uint64_t argc,
                             gab_value argv[argc]) {
   bool full = gab_chnisfull(gab_arg(0));
 
@@ -25,7 +25,7 @@ a_gab_value *gab_lib_isfull(struct gab_triple gab, size_t argc,
   return nullptr;
 }
 
-a_gab_value *gab_lib_isempty(struct gab_triple gab, size_t argc,
+a_gab_value *gab_lib_isempty(struct gab_triple gab, uint64_t argc,
                              gab_value argv[argc]) {
   bool empty = gab_chnisempty(gab_arg(0));
